Prepare Beta-regression Data

Create data frame of model data to be used in the beta-regression model for stray rates

prepare_stray_model_data(
  hatchery,
  type = c("natural", "hatchery"),
  sim_year,
  flow_oct_nov,
  flow_apr_may,
  releases,
  mean_PDO_return
)

Arguments

hatchery

name of hatchery to produce data for

type

either in river or bay release

sim_year

simulation year, used to calculate run year

flow_oct_nov

the mean flow for Oct-Nov, use DSMflow::hatchery_oct_nov_flows

flow_apr_may

the mean flow for Apr-May, use DSMflow::hatchery_apr_may_flows

mean_PDO_return

...

released

the total number of released hatchery fish

Details

Natural origin fish get a stray rate equal to Feather stray at distance = 0.