Spring Run Chinook Model

Spring Run Chinook life cycle model used for CVPIA's Structured Decision Making Process

spring_run_model(
  scenario = NULL,
  mode = c("seed", "simulate", "calibrate"),
  seeds = NULL,
  ..params = springRunDSM::params,
  stochastic = FALSE
)

Source

IP-117068

Arguments

scenario

Model inputs, can be modified to test management actions

mode

The mode to run model in. Can be "seed", "simulate", "calibrate"

seeds

The default value is NULL runs the model in seeding mode, returning a 31 by 25 matrix with the first four years of seeded adults. This returned value can be fed into the model again as the value for the seeds argument

..params

Parameters for model and submodels. Defaults to fallRunDSM::params.

stochastic

TRUE FALSE value indicating if model should be run stochastically. Defaults to FALSE.

Examples

spring_run_seeds <- springRunDSM::spring_run_model(mode = "seed")
springRunDSM::spring_run_model(scenario = DSMscenario::scenarios$ONE,
                           mode = "simulate",
                           seeds = spring_run_seeds)
#> $spawners
#>                               1    2    3    4    5    6    7    8    9   10
#> Upper Sacramento River       24   16    8   10   18   18   14   10   14    8
#> Antelope Creek               46   29   14   18   36   36   29   22   25   16
#> Battle Creek                519  330  169  255  450  419  281  287  326  191
#> Bear Creek                    0    0    0    0    0    0    0    0    0    0
#> Big Chico Creek              30   22   14   16   24   24   20   16   20   14
#> Butte Creek                5974 3856 1725 1678 2724 2353 1240 1494 1962 1022
#> Clear Creek                 535  345  172  416  813  811  670  828  947  568
#> Cottonwood Creek             24   16    8   10   18   18   14   10   14    8
#> Cow Creek                     0    0    0    0    0    0    0    0    0    0
#> Deer Creek                  347  213  100   96  152  144   98   78   84   60
#> Elder Creek                   0    0    0    0    0    0    0    0    0    0
#> Mill Creek                  406  242  100  118  208  177   90   75   82   46
#> Paynes Creek                  0    0    0    0    0    0    0    0    0    0
#> Stony Creek                   0    0    0    0    0    0    0    0    0    0
#> Thomes Creek                 24   16    8   10   18   18   14   10   14    8
#> Upper-mid Sacramento River    0    0    0    0    0    0    0    0    0    0
#> Sutter Bypass                 0    0    0    0    0    0    0    0    0    0
#> Bear River                    0    0    0    0    0    0    0    0    0    0
#> Feather River              5009 4996 5053 5646 6767 7840 7703 6495 5843 5554
#> Yuba River                  739  722  749  870  984 1000  931  939 1038  943
#> Lower-mid Sacramento River    0    0    0    0    0    0    0    0    0    0
#> Yolo Bypass                   0    0    0    0    0    0    0    0    0    0
#> American River                0    0    0    0    0    0    0    0    0    0
#> Lower Sacramento River        0    0    0    0    0    0    0    0    0    0
#> Calaveras River              24   16    8   10   18   18   14   10   14    8
#> Cosumnes River               24   16    8   10   18   18   14   10   14    8
#> Mokelumne River              24   16    8   10   18   18   14   10   14    8
#> Merced River                 24   16    8   10   18   18   14   10   14    8
#> Stanislaus River             24   16    8   10   18   18   14   10   14    8
#> Tuolumne River               24   16    8   10   18   18   14   10   14    8
#> San Joaquin River             0    0    0    0    0    0    0    0    0    0
#>                              11   12   13   14   15   16   17   18   19   20
#> Upper Sacramento River        8   10    8    2    8   10   14   20   18   16
#> Antelope Creek               14   22   14    6   14   22   25   38   36   30
#> Battle Creek                202  330  202   81  187  281  308  445  405  352
#> Bear Creek                    0    0    0    0    0    0    0    0    0    0
#> Big Chico Creek              14   16   14    8   14   16   20   26   24   22
#> Butte Creek                 936 1636  959  295  803 1215 1209 1458 1056  572
#> Clear Creek                 867 1456  857  311 1002 1585 1935 3255 3221 2792
#> Cottonwood Creek              8   10    8    2    8   10   14   20   18   16
#> Cow Creek                     0    0    0    0    0    0    0    0    0    0
#> Deer Creek                   56   66   54   38   56   68   76   98   88   82
#> Elder Creek                   0    0    0    0    0    0    0    0    0    0
#> Mill Creek                   40   70   40   12   29   50   62   86   78   68
#> Paynes Creek                  0    0    0    0    0    0    0    0    0    0
#> Stony Creek                   0    0    0    0    0    0    0    0    0    0
#> Thomes Creek                  8   10    8    2    8   10   14   20   18   16
#> Upper-mid Sacramento River    0    0    0    0    0    0    0    0    0    0
#> Sutter Bypass                 0    0    0    0    0    0    0    0    0    0
#> Bear River                    0    0    0    0    0    0    0    0    0    0
#> Feather River              5188 5095 5051 5034 5114 5230 5581 6151 6366 6419
#> Yuba River                  872  923  853  768  906 1073 1142 1346 1412 1442
#> Lower-mid Sacramento River    0    0    0    0    0    0    0    0    0    0
#> Yolo Bypass                   0    0    0    0    0    0    0    0    0    0
#> American River                0    0    0    0    0    0    0    0    0    0
#> Lower Sacramento River        0    0    0    0    0    0    0    0    0    0
#> Calaveras River               8   10    8    2    8   10   14   20   18   16
#> Cosumnes River                8   10    8    2    8   10   14   20   18   16
#> Mokelumne River               8   10    8    2    8   10   14   20   18   16
#> Merced River                  8   10    8    2    8   10   14   20   18   16
#> Stanislaus River              8   10    8    2    8   10   14   20   18   16
#> Tuolumne River                8   10    8    2    8   10   14   20   18   16
#> San Joaquin River             0    0    0    0    0    0    0    0    0    0
#> 
#> $juvenile_biomass
#>                                   1         2         3         4         5
#> Upper Sacramento River          0.0     628.0     628.0     628.0     628.0
#> Antelope Creek                  0.0    1130.4    1130.4    1287.4    1296.5
#> Battle Creek                10401.9  121152.8  176757.5  203589.4  211684.6
#> Bear Creek                      0.0       0.0       0.0       0.0       0.0
#> Big Chico Creek                 0.0       0.0       0.0       0.0       0.0
#> Butte Creek                 62448.0 1499041.8 2090587.3 2300666.3 2337929.0
#> Clear Creek                 19969.3  300794.4  445914.1  551453.8  578645.4
#> Cottonwood Creek                0.0       0.0       0.0       0.0       0.0
#> Cow Creek                       0.0       0.0       0.0       0.0       0.0
#> Deer Creek                   2339.5   68464.7   85880.3   91304.3   92216.2
#> Elder Creek                     0.0       0.0       0.0       0.0       0.0
#> Mill Creek                   3808.0   67755.2   88277.1   94563.0   96435.5
#> Paynes Creek                    0.0       0.0       0.0       0.0       0.0
#> Stony Creek                     0.0       0.0       0.0       0.0       0.0
#> Thomes Creek                    0.0       0.0       0.0       0.0       0.0
#> Upper-mid Sacramento River      0.0       0.0       0.0       0.0       0.0
#> Sutter Bypass                   0.0       0.0       0.0       0.0       0.0
#> Bear River                      0.0       0.0       0.0       0.0       0.0
#> Feather River              104542.1  913244.7 1845096.4 3153473.8 3232391.1
#> Yuba River                  48004.9  203948.9  357168.4  471240.3  492224.5
#> Lower-mid Sacramento River      0.0       0.0       0.0       0.0       0.0
#> Yolo Bypass                     0.0       0.0       0.0       0.0       0.0
#> American River                  0.0       0.0       0.0       0.0       0.0
#> Lower Sacramento River          0.0       0.0       0.0       0.0       0.0
#> Calaveras River                 0.0       0.0       0.0       0.0       0.0
#> Cosumnes River                  0.0       0.0       0.0       0.0       0.0
#> Mokelumne River                 0.0       0.0       0.0       0.0       0.0
#> Merced River                    0.0       0.0       0.0       0.0       0.0
#> Stanislaus River                0.0       0.0       0.0       0.0       0.0
#> Tuolumne River                  0.0       0.0       0.0       0.0       0.0
#> San Joaquin River               0.0       0.0       0.0       0.0       0.0
#>                                    6         7         8         9        10
#> Upper Sacramento River        1256.0    1256.0    1256.0    1256.0    1256.0
#> Antelope Creek                1641.9    1641.9    1641.9    1745.2    1745.2
#> Battle Creek                360456.8  363658.6  367504.2  401004.8  405195.4
#> Bear Creek                       0.0       0.0       0.0       0.0       0.0
#> Big Chico Creek                  0.0       0.0       0.0       0.0       0.0
#> Butte Creek                3227460.2 3235834.7 3258806.2 3443544.7 3463011.5
#> Clear Creek                1108632.1 1125203.2 1149578.4 1328275.3 1349364.9
#> Cottonwood Creek                 0.0       0.0       0.0       0.0       0.0
#> Cow Creek                        0.0       0.0       0.0       0.0       0.0
#> Deer Creek                  115986.9  116508.4  117076.3  119220.5  119656.9
#> Elder Creek                      0.0       0.0       0.0       0.0       0.0
#> Mill Creek                  128281.0  128986.8  129405.9  135317.5  135853.1
#> Paynes Creek                     0.0       0.0       0.0       0.0       0.0
#> Stony Creek                      0.0       0.0       0.0       0.0       0.0
#> Thomes Creek                     0.0       0.0       0.0       0.0       0.0
#> Upper-mid Sacramento River       0.0       0.0       0.0       0.0       0.0
#> Sutter Bypass                    0.0       0.0       0.0       0.0       0.0
#> Bear River                       0.0       0.0       0.0       0.0       0.0
#> Feather River              3910494.1 4009591.8 4076757.9 4095215.5 4109809.6
#> Yuba River                  740372.3  777402.8  812136.8  881092.5  914653.9
#> Lower-mid Sacramento River       0.0       0.0       0.0       0.0       0.0
#> Yolo Bypass                      0.0       0.0       0.0       0.0       0.0
#> American River                   0.0       0.0       0.0       0.0       0.0
#> Lower Sacramento River           0.0       0.0       0.0       0.0       0.0
#> Calaveras River                  0.0       0.0       0.0       0.0       0.0
#> Cosumnes River                   0.0       0.0       0.0       0.0       0.0
#> Mokelumne River                  0.0       0.0       0.0       0.0       0.0
#> Merced River                     0.0       0.0       0.0       0.0       0.0
#> Stanislaus River                 0.0       0.0       0.0       0.0       0.0
#> Tuolumne River                   0.0       0.0       0.0       0.0       0.0
#> San Joaquin River                0.0       0.0       0.0       0.0       0.0
#>                                   11        12        13        14        15
#> Upper Sacramento River        1256.0    1256.0    1256.0    1256.0    1256.0
#> Antelope Creek                1745.2    1808.0    1808.0    1808.0    1808.0
#> Battle Creek                408209.7  420543.4  479857.6  480537.9  555884.0
#> Bear Creek                       0.0       0.0       0.0       0.0       0.0
#> Big Chico Creek                  0.0       0.0       0.0       0.0       0.0
#> Butte Creek                3480957.6 3557219.3 3853388.0 3855387.5 4157886.0
#> Clear Creek                1371353.8 1469800.5 1905747.5 1913438.4 2557105.7
#> Cottonwood Creek                 0.0       0.0       0.0       0.0       0.0
#> Cow Creek                        0.0       0.0       0.0       0.0       0.0
#> Deer Creek                  119899.9  120730.4  127603.8  127896.4  136236.5
#> Elder Creek                      0.0       0.0       0.0       0.0       0.0
#> Mill Creek                  136217.6  137348.9  145337.2  145337.2  155755.6
#> Paynes Creek                     0.0       0.0       0.0       0.0       0.0
#> Stony Creek                      0.0       0.0       0.0       0.0       0.0
#> Thomes Creek                     0.0       0.0       0.0       0.0       0.0
#> Upper-mid Sacramento River       0.0       0.0       0.0       0.0       0.0
#> Sutter Bypass                    0.0       0.0       0.0       0.0       0.0
#> Bear River                       0.0       0.0       0.0       0.0       0.0
#> Feather River              4125445.8 4151137.8 4272826.7 4326705.8 4922605.4
#> Yuba River                  928549.9  956306.1 1233575.9 1258536.1 1544612.2
#> Lower-mid Sacramento River       0.0       0.0       0.0       0.0       0.0
#> Yolo Bypass                      0.0       0.0       0.0       0.0       0.0
#> American River                   0.0       0.0       0.0       0.0       0.0
#> Lower Sacramento River           0.0       0.0       0.0       0.0       0.0
#> Calaveras River                  0.0       0.0       0.0       0.0       0.0
#> Cosumnes River                   0.0       0.0       0.0       0.0       0.0
#> Mokelumne River                  0.0       0.0       0.0       0.0       0.0
#> Merced River                     0.0       0.0       0.0       0.0       0.0
#> Stanislaus River                 0.0       0.0       0.0       0.0       0.0
#> Tuolumne River                   0.0       0.0       0.0       0.0       0.0
#> San Joaquin River                0.0       0.0       0.0       0.0       0.0
#>                                   16        17        18        19        20
#> Upper Sacramento River        1256.0    1256.0    1955.9    2740.9    3400.3
#> Antelope Creek                2162.5    2225.3    2548.4    2862.4    3145.0
#> Battle Creek                641766.5  658244.1  784756.1  917066.0 1021953.7
#> Bear Creek                       0.0       0.0       0.0       0.0       0.0
#> Big Chico Creek                  0.0       0.0       0.0       0.0       0.0
#> Butte Creek                4414862.7 4459758.2 4711814.3 5033299.7 5193776.8
#> Clear Creek                3503769.1 3687814.5 4800244.5 6098009.6 7171367.8
#> Cottonwood Creek                 0.0       0.0       0.0       0.0       0.0
#> Cow Creek                        0.0       0.0       0.0       0.0       0.0
#> Deer Creek                  143371.1  146857.4  160796.7  175084.6  187959.5
#> Elder Creek                      0.0       0.0       0.0       0.0       0.0
#> Mill Creek                  165457.7  167100.5  184855.1  203826.6  220891.8
#> Paynes Creek                     0.0       0.0       0.0       0.0       0.0
#> Stony Creek                      0.0       0.0       0.0       0.0       0.0
#> Thomes Creek                     0.0       0.0       0.0       0.0       0.0
#> Upper-mid Sacramento River       0.0       0.0       0.0       0.0       0.0
#> Sutter Bypass                    0.0       0.0       0.0       0.0       0.0
#> Bear River                       0.0       0.0       0.0       0.0       0.0
#> Feather River              5618501.4 6152425.5 6825257.7 9291537.5 9983634.6
#> Yuba River                 1938870.0 2129240.6 2459146.2 2871062.8 3289979.1
#> Lower-mid Sacramento River       0.0       0.0       0.0       0.0       0.0
#> Yolo Bypass                      0.0       0.0       0.0       0.0       0.0
#> American River                   0.0       0.0       0.0       0.0       0.0
#> Lower Sacramento River           0.0       0.0       0.0       0.0       0.0
#> Calaveras River                  0.0       0.0       0.0       0.0       0.0
#> Cosumnes River                   0.0       0.0       0.0       0.0       0.0
#> Mokelumne River                  0.0       0.0       0.0       0.0       0.0
#> Merced River                     0.0       0.0       0.0       0.0       0.0
#> Stanislaus River                 0.0       0.0       0.0     502.4     502.4
#> Tuolumne River                   0.0       0.0       0.0       0.0       0.0
#> San Joaquin River                0.0       0.0       0.0       0.0       0.0
#> 
#> $proportion_natural
#>                                      1           2          3         4
#> Upper Sacramento River     0.999995833 0.999993750 0.99998750 0.9999900
#> Antelope Creek             0.999997826 0.999996552 0.99999286 0.9999944
#> Battle Creek               0.951830260 0.924242144 0.85207050 0.9019604
#> Bear Creek                 0.000000000 0.000000000 0.00000000 0.0000000
#> Big Chico Creek            0.799997333 0.727269422 0.57142449 0.6249961
#> Butte Creek                0.997321711 0.995850597 0.99072458 0.9904648
#> Clear Creek                0.996261496 0.994202610 0.98837152 0.9951921
#> Cottonwood Creek           0.999995833 0.999993750 0.99998750 0.9999900
#> Cow Creek                  0.000000000 0.000000000 0.00000000 0.0000000
#> Deer Creek                 0.913544405 0.859154526 0.69999930 0.6874993
#> Elder Creek                0.000000000 0.000000000 0.00000000 0.0000000
#> Mill Creek                 0.999999754 0.999999587 0.99999900 0.9999992
#> Paynes Creek               0.000000000 0.000000000 0.00000000 0.0000000
#> Stony Creek                0.000000000 0.000000000 0.00000000 0.0000000
#> Thomes Creek               0.999995833 0.999993750 0.99998750 0.9999900
#> Upper-mid Sacramento River 0.000000000 0.000000000 0.00000000 0.0000000
#> Sutter Bypass              0.000000000 0.000000000 0.00000000 0.0000000
#> Bear River                 0.000000000 0.000000000 0.00000000 0.0000000
#> Feather River              0.007187063 0.004603683 0.01583218 0.1191994
#> Yuba River                 0.062246270 0.040166199 0.07476635 0.2034483
#> Lower-mid Sacramento River 0.000000000 0.000000000 0.00000000 0.0000000
#> Yolo Bypass                0.000000000 0.000000000 0.00000000 0.0000000
#> American River             0.000000000 0.000000000 0.00000000 0.0000000
#> Lower Sacramento River     0.000000000 0.000000000 0.00000000 0.0000000
#> Calaveras River            0.999995833 0.999993750 0.99998750 0.9999900
#> Cosumnes River             0.999995833 0.999993750 0.99998750 0.9999900
#> Mokelumne River            0.999995833 0.999993750 0.99998750 0.9999900
#> Merced River               0.999995833 0.999993750 0.99998750 0.9999900
#> Stanislaus River           0.999995833 0.999993750 0.99998750 0.9999900
#> Tuolumne River             0.999995833 0.999993750 0.99998750 0.9999900
#> San Joaquin River          0.000000000 0.000000000 0.00000000 0.0000000
#>                                    5         6         7         8         9
#> Upper Sacramento River     0.9999944 0.9999944 0.9999929 0.9999900 0.9999929
#> Antelope Creek             0.9999972 0.9999972 0.9999966 0.9999955 0.9999960
#> Battle Creek               0.9444442 0.9403339 0.9110317 0.9128917 0.9233126
#> Bear Creek                 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Big Chico Creek            0.7499969 0.7499969 0.6999965 0.6249961 0.6999965
#> Butte Creek                0.9941262 0.9932001 0.9870967 0.9892904 0.9918450
#> Clear Creek                0.9975399 0.9975338 0.9970148 0.9975844 0.9978880
#> Cottonwood Creek           0.9999944 0.9999944 0.9999929 0.9999900 0.9999929
#> Cow Creek                  0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Deer Creek                 0.8026311 0.7916661 0.6938768 0.6153838 0.6428564
#> Elder Creek                0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Mill Creek                 0.9999995 0.9999994 0.9999989 0.9999987 0.9999988
#> Paynes Creek               0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Stony Creek                0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Thomes Creek               0.9999944 0.9999944 0.9999929 0.9999900 0.9999929
#> Upper-mid Sacramento River 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Sutter Bypass              0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Bear River                 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Feather River              0.2651101 0.3656888 0.3544074 0.2343341 0.1488961
#> Yuba River                 0.2957317 0.3070000 0.2556391 0.2619808 0.3323699
#> Lower-mid Sacramento River 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Yolo Bypass                0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> American River             0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Lower Sacramento River     0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Calaveras River            0.9999944 0.9999944 0.9999929 0.9999900 0.9999929
#> Cosumnes River             0.9999944 0.9999944 0.9999929 0.9999900 0.9999929
#> Mokelumne River            0.9999944 0.9999944 0.9999929 0.9999900 0.9999929
#> Merced River               0.9999944 0.9999944 0.9999929 0.9999900 0.9999929
#> Stanislaus River           0.9999944 0.9999944 0.9999929 0.9999900 0.9999929
#> Tuolumne River             0.9999944 0.9999944 0.9999929 0.9999900 0.9999929
#> San Joaquin River          0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>                                   10         11         12         13
#> Upper Sacramento River     0.9999875 0.99998750 0.99999000 0.99998750
#> Antelope Creek             0.9999938 0.99999286 0.99999545 0.99999286
#> Battle Creek               0.8691095 0.87623719 0.92424214 0.87623719
#> Bear Creek                 0.0000000 0.00000000 0.00000000 0.00000000
#> Big Chico Creek            0.5714245 0.57142449 0.62499609 0.57142449
#> Butte Creek                0.9843443 0.98290588 0.99021999 0.98331585
#> Clear Creek                0.9964787 0.99769308 0.99862631 0.99766616
#> Cottonwood Creek           0.9999875 0.99998750 0.99999000 0.99998750
#> Cow Creek                  0.0000000 0.00000000 0.00000000 0.00000000
#> Deer Creek                 0.4999992 0.46428489 0.54545372 0.44444362
#> Elder Creek                0.0000000 0.00000000 0.00000000 0.00000000
#> Mill Creek                 0.9999978 0.99999750 0.99999857 0.99999750
#> Paynes Creek               0.0000000 0.00000000 0.00000000 0.00000000
#> Stony Creek                0.0000000 0.00000000 0.00000000 0.00000000
#> Thomes Creek               0.9999875 0.99998750 0.99999000 0.99998750
#> Upper-mid Sacramento River 0.0000000 0.00000000 0.00000000 0.00000000
#> Sutter Bypass              0.0000000 0.00000000 0.00000000 0.00000000
#> Bear River                 0.0000000 0.00000000 0.00000000 0.00000000
#> Feather River              0.1046093 0.04144179 0.02394504 0.01544249
#> Yuba River                 0.2651113 0.20527521 0.24918741 0.18757325
#> Lower-mid Sacramento River 0.0000000 0.00000000 0.00000000 0.00000000
#> Yolo Bypass                0.0000000 0.00000000 0.00000000 0.00000000
#> American River             0.0000000 0.00000000 0.00000000 0.00000000
#> Lower Sacramento River     0.0000000 0.00000000 0.00000000 0.00000000
#> Calaveras River            0.9999875 0.99998750 0.99999000 0.99998750
#> Cosumnes River             0.9999875 0.99998750 0.99999000 0.99998750
#> Mokelumne River            0.9999875 0.99998750 0.99999000 0.99998750
#> Merced River               0.9999875 0.99998750 0.99999000 0.99998750
#> Stanislaus River           0.9999875 0.99998750 0.99999000 0.99998750
#> Tuolumne River             0.9999875 0.99998750 0.99999000 0.99998750
#> San Joaquin River          0.0000000 0.00000000 0.00000000 0.00000000
#>                                    14         15         16        17        18
#> Upper Sacramento River     0.99995000 0.99998750 0.99999000 0.9999929 0.9999950
#> Antelope Creek             0.99998333 0.99999286 0.99999545 0.9999960 0.9999974
#> Battle Creek               0.69135717 0.86630970 0.91103170 0.9188309 0.9438200
#> Bear Creek                 0.00000000 0.00000000 0.00000000 0.0000000 0.0000000
#> Big Chico Creek            0.24999688 0.57142449 0.62499609 0.6999965 0.7692278
#> Butte Creek                0.94576239 0.98007460 0.98683119 0.9867658 0.9890260
#> Clear Creek                0.99356881 0.99800389 0.99873811 0.9989664 0.9993855
#> Cottonwood Creek           0.99995000 0.99998750 0.99999000 0.9999929 0.9999950
#> Cow Creek                  0.00000000 0.00000000 0.00000000 0.0000000 0.0000000
#> Deer Creek                 0.21052576 0.46428489 0.55882271 0.6052624 0.6938768
#> Elder Creek                0.00000000 0.00000000 0.00000000 0.0000000 0.0000000
#> Mill Creek                 0.99999167 0.99999655 0.99999800 0.9999984 0.9999988
#> Paynes Creek               0.00000000 0.00000000 0.00000000 0.0000000 0.0000000
#> Stony Creek                0.00000000 0.00000000 0.00000000 0.0000000 0.0000000
#> Thomes Creek               0.99995000 0.99998750 0.99999000 0.9999929 0.9999950
#> Upper-mid Sacramento River 0.00000000 0.00000000 0.00000000 0.0000000 0.0000000
#> Sutter Bypass              0.00000000 0.00000000 0.00000000 0.0000000 0.0000000
#> Bear River                 0.00000000 0.00000000 0.00000000 0.0000000 0.0000000
#> Feather River              0.01211760 0.02757137 0.04913958 0.1089410 0.1915136
#> Yuba River                 0.09765624 0.23509931 0.35414722 0.3931698 0.4851411
#> Lower-mid Sacramento River 0.00000000 0.00000000 0.00000000 0.0000000 0.0000000
#> Yolo Bypass                0.00000000 0.00000000 0.00000000 0.0000000 0.0000000
#> American River             0.00000000 0.00000000 0.00000000 0.0000000 0.0000000
#> Lower Sacramento River     0.00000000 0.00000000 0.00000000 0.0000000 0.0000000
#> Calaveras River            0.99995000 0.99998750 0.99999000 0.9999929 0.9999950
#> Cosumnes River             0.99995000 0.99998750 0.99999000 0.9999929 0.9999950
#> Mokelumne River            0.99995000 0.99998750 0.99999000 0.9999929 0.9999950
#> Merced River               0.99995000 0.99998750 0.99999000 0.9999929 0.9999950
#> Stanislaus River           0.99995000 0.99998750 0.99999000 0.9999929 0.9999950
#> Tuolumne River             0.99995000 0.99998750 0.99999000 0.9999929 0.9999950
#> San Joaquin River          0.00000000 0.00000000 0.00000000 0.0000000 0.0000000
#>                                   19        20
#> Upper Sacramento River     0.9999944 0.9999938
#> Antelope Creek             0.9999972 0.9999967
#> Battle Creek               0.9382714 0.9289770
#> Bear Creek                 0.0000000 0.0000000
#> Big Chico Creek            0.7499969 0.7272694
#> Butte Creek                0.9848484 0.9720278
#> Clear Creek                0.9993790 0.9992836
#> Cottonwood Creek           0.9999944 0.9999938
#> Cow Creek                  0.0000000 0.0000000
#> Deer Creek                 0.6590902 0.6341456
#> Elder Creek                0.0000000 0.0000000
#> Mill Creek                 0.9999987 0.9999985
#> Paynes Creek               0.0000000 0.0000000
#> Stony Creek                0.0000000 0.0000000
#> Thomes Creek               0.9999944 0.9999938
#> Upper-mid Sacramento River 0.0000000 0.0000000
#> Sutter Bypass              0.0000000 0.0000000
#> Bear River                 0.0000000 0.0000000
#> Feather River              0.2188187 0.2252687
#> Yuba River                 0.5092068 0.5194174
#> Lower-mid Sacramento River 0.0000000 0.0000000
#> Yolo Bypass                0.0000000 0.0000000
#> American River             0.0000000 0.0000000
#> Lower Sacramento River     0.0000000 0.0000000
#> Calaveras River            0.9999944 0.9999938
#> Cosumnes River             0.9999944 0.9999938
#> Mokelumne River            0.9999944 0.9999938
#> Merced River               0.9999944 0.9999938
#> Stanislaus River           0.9999944 0.9999938
#> Tuolumne River             0.9999944 0.9999938
#> San Joaquin River          0.0000000 0.0000000
#> 
#> $viability_metrics
#>       [,1] [,2] [,3] [,4]
#>  [1,]    0    0    1    0
#>  [2,]    0    0    1    0
#>  [3,]    0    0    1    0
#>  [4,]    0    0    0    0
#>  [5,]    0    0    1    0
#>  [6,]    0    0    1    0
#>  [7,]    0    0    0    0
#>  [8,]    0    0    0    0
#>  [9,]    0    0    1    0
#> [10,]    0    0    1    0
#> [11,]    0    0    0    0
#> [12,]    1    0    1    0
#> [13,]    0    0    0    0
#> [14,]    0    0    0    0
#> [15,]    0    0    0    0
#> [16,]    0    0    1    0
#> [17,]    0    0    0    0
#> [18,]    1    0    1    0
#> [19,]    1    0    0    0
#>