Spring Run Chinook Model

Spring Run Chinook life cycle model used for CVPIA's Structured Decision Making Process

spring_run_model(
  scenario = NULL,
  mode = c("seed", "simulate", "calibrate"),
  seeds = NULL,
  ..params = springRunDSM::params
)

Arguments

scenario
Model inputs, can be modified to test management actions

mode
The mode to run model in. Can be "seed", "simulate", "calibrate"

seeds
The default value is NULL runs the model in seeding mode, returning a 31 by 25 matrix with the first four years of seeded adults. This returned value can be fed into the model again as the value for the seeds argument

..params
Parameters for model and submodels

Source

IP-117068

Examples

spring_run_seeds <- springRunDSM::spring_run_model(mode = "seed")
springRunDSM::spring_run_model(scenario = DSMscenario::scenarios$ONE,
                           mode = "simulate",
                           seeds = spring_run_seeds)

#> $spawners
#>                                1     2    3    4     5     6     7    8    9
#> Upper Sacramento River      2143  1168  433  508  1003  1355  1443  881  456
#> Antelope Creek               606   347  180  204   300   379   340  225   91
#> Battle Creek                1664  1282  493  160   426   907  1216  625  252
#> Bear Creek                     0     0    0    0     0     0     0    0    0
#> Big Chico Creek              445   274  173  221   298   364   322  189   90
#> Butte Creek                 7026  3983 2209 2778  4131  4453  4290 2849 1642
#> Clear Creek                  835   628  235   97   246   447   605  299  124
#> Cottonwood Creek             123    81   34   20    40    77    77   25    8
#> Cow Creek                      0     0    0    0     0     0     0    0    0
#> Deer Creek                  1141   784  386  342   478   676   757  457  176
#> Elder Creek                    0     0    0    0     0     0     0    0    0
#> Mill Creek                  1577  1179  562  416   722  1035  1040  681  253
#> Paynes Creek                   0     0    0    0     0     0     0    0    0
#> Stony Creek                    0     0    0    0     0     0     0    0    0
#> Thomes Creek                 120    61   26   24    42    49    65   32   14
#> Upper-mid Sacramento River     0     0    0    0     0     0     0    0    0
#> Sutter Bypass                  0     0    0    0     0     0     0    0    0
#> Bear River                     0     0    0    0     0     0     0    0    0
#> Feather River              31687 26249 9901 8335 14762 20734 16049 9141 6997
#> Yuba River                  1358  1330 1036 1014  1195  1267   937  853  987
#> Lower-mid Sacramento River     0     0    0    0     0     0     0    0    0
#> Yolo Bypass                    0     0    0    0     0     0     0    0    0
#> American River                 0     0    0    0     0     0     0    0    0
#> Lower Sacramento River         0     0    0    0     0     0     0    0    0
#> Calaveras River               91    59   22   20    46    47    41   18   12
#> Cosumnes River               104    62   31   20    40    46    55   22   13
#> Mokelumne River               92    54   22   25    46    43    56   24   11
#> Merced River                  88    57   24   27    36    53    57   21   16
#> Stanislaus River              90    68   26   30    41    57    50   32   16
#> Tuolumne River               103    60   29   22    49    70    48   24   10
#> San Joaquin River              0     0    0    0     0     0     0    0    0
#>                              10   11   12   13   14   15   16   17   18   19
#> Upper Sacramento River      286  154  159  114   44  219  445  627 1561 2139
#> Antelope Creek               22   20    7   22    8   22   22   54  127  264
#> Battle Creek                151   86   85   64   38  147  264  555 1129 1414
#> Bear Creek                    0    0    0    0    0    0    0    0    0    0
#> Big Chico Creek              29   19   27    8   12    4   18   37  107  255
#> Butte Creek                1003  437  401  316  197  593 1141 2025 5210 7739
#> Clear Creek                  55   31   33   17    2   22   42  101  267  351
#> Cottonwood Creek             10    3    4    2    2    2   14   10   25   43
#> Cow Creek                     0    0    0    0    0    0    0    0    0    0
#> Deer Creek                  108   83   91   70   45  113  147  304  643  907
#> Elder Creek                   0    0    0    0    0    0    0    0    0    0
#> Mill Creek                  152   68   79   44   25   77  159  294  767 1117
#> Paynes Creek                  0    0    0    0    0    0    0    0    0    0
#> Stony Creek                   0    0    0    0    0    0    0    0    0    0
#> Thomes Creek                  7    0    0    1    2    7    4   15   37   31
#> Upper-mid Sacramento River    0    0    0    0    0    0    0    0    0    0
#> Sutter Bypass                 0    0    0    0    0    0    0    0    0    0
#> Bear River                    0    0    0    0    0    0    0    0    0    0
#> Feather River              5596 7140 4523 4422 3873 5349 5270 4318 4303 5724
#> Yuba River                  730 1018  692  653  546  749  737  618  726 1067
#> Lower-mid Sacramento River    0    0    0    0    0    0    0    0    0    0
#> Yolo Bypass                   0    0    0    0    0    0    0    0    0    0
#> American River                0    0    0    0    0    0    0    0    0    0
#> Lower Sacramento River        0    0    0    0    0    0    0    0    0    0
#> Calaveras River               5    3    3    1    0    2    1   17   28   33
#> Cosumnes River                5    3    1    2    2    5    5   10   18   43
#> Mokelumne River               2    2    1    3    1    1    8   12   25   41
#> Merced River                  8    3    1    0    1    4    5   10   21   35
#> Stanislaus River             11    3    2    2    1    2    4   12   30   35
#> Tuolumne River                9    4    1    0    2    3    4   10   18   37
#> San Joaquin River             0    0    0    0    0    0    0    0    0    0
#>                              20
#> Upper Sacramento River     2248
#> Antelope Creek              467
#> Battle Creek               1397
#> Bear Creek                    0
#> Big Chico Creek             452
#> Butte Creek                9333
#> Clear Creek                 398
#> Cottonwood Creek             58
#> Cow Creek                     0
#> Deer Creek                 1299
#> Elder Creek                   0
#> Mill Creek                 1470
#> Paynes Creek                  0
#> Stony Creek                   0
#> Thomes Creek                 50
#> Upper-mid Sacramento River    0
#> Sutter Bypass                 0
#> Bear River                    0
#> Feather River              5244
#> Yuba River                 1207
#> Lower-mid Sacramento River    0
#> Yolo Bypass                   0
#> American River                0
#> Lower Sacramento River        0
#> Calaveras River              38
#> Cosumnes River               42
#> Mokelumne River              41
#> Merced River                 37
#> Stanislaus River             40
#> Tuolumne River               49
#> San Joaquin River             0
#> 
#> $natural_spawners
#>                                1     2    3    4     5     6     7    8    9
#> Upper Sacramento River      2143  1168  433  508  1003  1355  1443  881  456
#> Antelope Creek               605   345  179  204   299   379   340  225   90
#> Battle Creek                1643  1240  465  137   399   879  1189  606  220
#> Bear Creek                     0     0    0    0     0     0     0    0    0
#> Big Chico Creek              442   270  169  215   294   359   315  187   83
#> Butte Creek                 7008  3963 2192 2761  4119  4433  4274 2838 1624
#> Clear Creek                  833   625  234   95   246   443   604  299  116
#> Cottonwood Creek             122    81   34   19    39    77    76   23    8
#> Cow Creek                      0     0    0    0     0     0     0    0    0
#> Deer Creek                  1114   750  347  320   451   633   728  436  153
#> Elder Creek                    0     0    0    0     0     0     0    0    0
#> Mill Creek                  1577  1179  562  416   722  1035  1040  681  253
#> Paynes Creek                   0     0    0    0     0     0     0    0    0
#> Stony Creek                    0     0    0    0     0     0     0    0    0
#> Thomes Creek                 120    61   26   24    42    49    65   32   14
#> Upper-mid Sacramento River     0     0    0    0     0     0     0    0    0
#> Sutter Bypass                  0     0    0    0     0     0     0    0    0
#> Bear River                     0     0    0    0     0     0     0    0    0
#> Feather River              27115 19764 4310 3461 10270 14699 11858 4980 1030
#> Yuba River                   720   435  269  343   519   422   372  289  151
#> Lower-mid Sacramento River     0     0    0    0     0     0     0    0    0
#> Yolo Bypass                    0     0    0    0     0     0     0    0    0
#> American River                 0     0    0    0     0     0     0    0    0
#> Lower Sacramento River         0     0    0    0     0     0     0    0    0
#> Calaveras River               91    59   22   20    46    47    41   18   12
#> Cosumnes River               104    62   31   20    40    46    55   22   13
#> Mokelumne River               92    54   22   25    46    43    56   24   11
#> Merced River                  88    57   24   27    36    53    57   21   16
#> Stanislaus River              90    68   26   30    41    57    50   32   16
#> Tuolumne River               103    60   29   22    49    70    48   24   10
#> San Joaquin River              0     0    0    0     0     0     0    0    0
#>                             10  11  12  13  14  15   16   17   18   19   20
#> Upper Sacramento River     286 154 159 114  44 219  445  627 1561 2139 2248
#> Antelope Creek              22  20   7  22   7  21   21   54  126  264  466
#> Battle Creek               120  51  61  37  16 118  233  536 1113 1380 1370
#> Bear Creek                   0   0   0   0   0   0    0    0    0    0    0
#> Big Chico Creek             26  13  22   6   9   4   15   34  104  252  450
#> Butte Creek                986 420 386 304 182 581 1119 2014 5203 7715 9319
#> Clear Creek                 53  29  32  14   2  19   40   98  262  351  396
#> Cottonwood Creek             8   3   2   2   1   1   13   10   25   43   57
#> Cow Creek                    0   0   0   0   0   0    0    0    0    0    0
#> Deer Creek                  70  46  55  40  13  74  120  283  622  879 1267
#> Elder Creek                  0   0   0   0   0   0    0    0    0    0    0
#> Mill Creek                 152  68  79  44  25  77  159  294  767 1117 1470
#> Paynes Creek                 0   0   0   0   0   0    0    0    0    0    0
#> Stony Creek                  0   0   0   0   0   0    0    0    0    0    0
#> Thomes Creek                 7   0   0   1   2   7    4   15   37   31   50
#> Upper-mid Sacramento River   0   0   0   0   0   0    0    0    0    0    0
#> Sutter Bypass                0   0   0   0   0   0    0    0    0    0    0
#> Bear River                   0   0   0   0   0   0    0    0    0    0    0
#> Feather River              514  16  18  12  11 126  262  175  141  123   88
#> Yuba River                  79  48  33  25  10  16   23   43  127  247  498
#> Lower-mid Sacramento River   0   0   0   0   0   0    0    0    0    0    0
#> Yolo Bypass                  0   0   0   0   0   0    0    0    0    0    0
#> American River               0   0   0   0   0   0    0    0    0    0    0
#> Lower Sacramento River       0   0   0   0   0   0    0    0    0    0    0
#> Calaveras River              5   3   3   1   0   2    1   17   28   33   38
#> Cosumnes River               5   3   1   2   2   5    5   10   18   43   42
#> Mokelumne River              2   2   1   3   1   1    8   12   25   41   41
#> Merced River                 8   3   1   0   1   4    5   10   21   35   37
#> Stanislaus River            11   3   2   2   1   2    4   12   30   35   40
#> Tuolumne River               9   4   1   0   2   3    4   10   18   37   49
#> San Joaquin River            0   0   0   0   0   0    0    0    0    0    0
#> 
#> $juvenile_biomass
#>                                    1         2          3          4          5
#> Upper Sacramento River     1692020.4 4677084.5  6258842.2  7572361.1  7607349.2
#> Antelope Creek              616005.2 1428040.6  1579639.8  2058081.6  2109734.6
#> Battle Creek                  3893.6  451155.2  1141201.6  1615059.0  1624479.0
#> Bear Creek                       0.0       0.0        0.0        0.0        0.0
#> Big Chico Creek             667093.0 1523810.6  1648405.8  2132970.6  2189553.4
#> Butte Creek                3644409.6 7629917.4  9005143.2 11360739.8 11566001.6
#> Clear Creek                  14758.0  239310.3   510095.7   677651.8   680477.8
#> Cottonwood Creek               188.4   51370.4   102615.2   155681.2   156654.6
#> Cow Creek                        0.0       0.0        0.0        0.0        0.0
#> Deer Creek                  427197.0  757619.2   958956.0  1302911.6  1327152.4
#> Elder Creek                      0.0       0.0        0.0        0.0        0.0
#> Mill Creek                 1035477.8 2847383.4  3540507.0  4645849.8  4744194.6
#> Paynes Creek                     0.0       0.0        0.0        0.0        0.0
#> Stony Creek                      0.0       0.0        0.0        0.0        0.0
#> Thomes Creek                     0.0    2041.0     2386.4     2920.2     2920.2
#> Upper-mid Sacramento River       0.0       0.0        0.0        0.0        0.0
#> Sutter Bypass                    0.0       0.0        0.0        0.0        0.0
#> Bear River                       0.0       0.0        0.0        0.0        0.0
#> Feather River                  565.2 6366098.8 10750323.8 14867460.4 14898829.0
#> Yuba River                  980056.8 1751868.8  1832535.4  2215772.4  2336725.2
#> Lower-mid Sacramento River       0.0       0.0        0.0        0.0        0.0
#> Yolo Bypass                      0.0       0.0        0.0        0.0        0.0
#> American River                   0.0       0.0        0.0        0.0        0.0
#> Lower Sacramento River           0.0       0.0        0.0        0.0        0.0
#> Calaveras River                  0.0       0.0        0.0        0.0        0.0
#> Cosumnes River                   0.0       0.0        0.0        0.0        0.0
#> Mokelumne River                  0.0       0.0        0.0        0.0        0.0
#> Merced River                     0.0       0.0        0.0        0.0        0.0
#> Stanislaus River                 0.0       0.0     7818.6     7818.6     7818.6
#> Tuolumne River                   0.0       0.0      345.4      376.8      376.8
#> San Joaquin River                0.0       0.0        0.0        0.0        0.0
#>                                     6          7          8          9
#> Upper Sacramento River      8384570.4  8479518.7  8504033.0  8650072.0
#> Antelope Creek              2141228.8  2170054.0  2183618.8  2204688.2
#> Battle Creek                1800821.4  1810210.0  1813381.4  1842834.6
#> Bear Creek                        0.0        0.0        0.0        0.0
#> Big Chico Creek             2224030.6  2256937.8  2271978.4  2293110.6
#> Butte Creek                12182791.8 12415999.6 12475659.6 12596581.0
#> Clear Creek                  758601.9   763098.0   765554.0   778722.4
#> Cottonwood Creek             156654.6   156654.6   156874.4   157062.8
#> Cow Creek                         0.0        0.0        0.0        0.0
#> Deer Creek                  1371520.6  1392119.0  1403014.8  1418369.4
#> Elder Creek                       0.0        0.0        0.0        0.0
#> Mill Creek                  4995802.8  5055651.2  5085198.6  5155377.6
#> Paynes Creek                      0.0        0.0        0.0        0.0
#> Stony Creek                       0.0        0.0        0.0        0.0
#> Thomes Creek                   3328.4     3328.4     3359.8     3359.8
#> Upper-mid Sacramento River        0.0        0.0        0.0        0.0
#> Sutter Bypass                     0.0        0.0        0.0        0.0
#> Bear River                        0.0        0.0        0.0        0.0
#> Feather River              15112506.0 15116462.4 15116713.6 15117812.6
#> Yuba River                  2403544.4  2457960.6  2474759.6  2494918.4
#> Lower-mid Sacramento River        0.0        0.0        0.0        0.0
#> Yolo Bypass                       0.0        0.0        0.0        0.0
#> American River                    0.0        0.0        0.0        0.0
#> Lower Sacramento River            0.0        0.0        0.0        0.0
#> Calaveras River                   0.0        0.0        0.0        0.0
#> Cosumnes River                    0.0        0.0        0.0        0.0
#> Mokelumne River                   0.0        0.0        0.0        0.0
#> Merced River                      0.0        0.0        0.0        0.0
#> Stanislaus River               7818.6     7818.6     7818.6     7818.6
#> Tuolumne River                  376.8      376.8      376.8      376.8
#> San Joaquin River                 0.0        0.0        0.0        0.0
#>                                    10         11         12         13
#> Upper Sacramento River      8696287.9  8718707.5  8817367.5 11025038.7
#> Antelope Creek              2227390.4  2228740.6  2251442.8  2326645.8
#> Battle Creek                1847481.8  1854546.8  1873418.2  2448132.4
#> Bear Creek                        0.0        0.0        0.0        0.0
#> Big Chico Creek             2318199.2  2319392.4  2342314.4  2388221.2
#> Butte Creek                12765889.8 12775561.0 12958089.2 14066320.8
#> Clear Creek                  780494.9   781791.4   785663.6   854729.9
#> Cottonwood Creek             157094.2   157094.2   157188.4   178666.0
#> Cow Creek                         0.0        0.0        0.0        0.0
#> Deer Creek                  1434352.0  1435356.8  1452124.4  1608873.2
#> Elder Creek                       0.0        0.0        0.0        0.0
#> Mill Creek                  5197045.4  5203953.4  5255198.2  5829378.6
#> Paynes Creek                      0.0        0.0        0.0        0.0
#> Stony Creek                       0.0        0.0        0.0        0.0
#> Thomes Creek                   3359.8     3359.8     3359.8     3673.8
#> Upper-mid Sacramento River        0.0        0.0        0.0        0.0
#> Sutter Bypass                     0.0        0.0        0.0        0.0
#> Bear River                        0.0        0.0        0.0        0.0
#> Feather River              15119947.8 15119947.8 15122365.6 15291046.4
#> Yuba River                  2515956.4  2517338.0  2537496.8  2557435.8
#> Lower-mid Sacramento River        0.0        0.0        0.0        0.0
#> Yolo Bypass                       0.0        0.0        0.0        0.0
#> American River                    0.0        0.0        0.0        0.0
#> Lower Sacramento River            0.0        0.0        0.0        0.0
#> Calaveras River                   0.0        0.0        0.0        0.0
#> Cosumnes River                    0.0        0.0        0.0        0.0
#> Mokelumne River                   0.0        0.0        0.0        0.0
#> Merced River                      0.0        0.0        0.0        0.0
#> Stanislaus River               7818.6     7818.6     7818.6     7818.6
#> Tuolumne River                  376.8      376.8      376.8      376.8
#> San Joaquin River                 0.0        0.0        0.0        0.0
#>                                    14         15         16         17
#> Upper Sacramento River     11098699.7 12871059.7 14924516.1 15546252.5
#> Antelope Creek              2357041.0  2385018.4  3037604.6  3687427.6
#> Battle Creek                2452465.6  3263998.6  3755377.2  4012449.0
#> Bear Creek                        0.0        0.0        0.0        0.0
#> Big Chico Creek             2421662.2  2481008.2  3219599.0  3927135.2
#> Butte Creek                14291992.6 15745530.0 21315827.2 24344294.4
#> Clear Creek                  855232.3   999609.5  1109152.1  1155658.5
#> Cottonwood Creek             178697.4   185479.8   198353.8   206203.8
#> Cow Creek                         0.0        0.0        0.0        0.0
#> Deer Creek                  1629189.0  1843776.6  2330822.0  2805370.2
#> Elder Creek                       0.0        0.0        0.0        0.0
#> Mill Creek                  5882790.0  6581251.6  8031021.0  9363542.8
#> Paynes Creek                      0.0        0.0        0.0        0.0
#> Stony Creek                       0.0        0.0        0.0        0.0
#> Thomes Creek                   3673.8     3862.2     3862.2     4239.0
#> Upper-mid Sacramento River        0.0        0.0        0.0        0.0
#> Sutter Bypass                     0.0        0.0        0.0        0.0
#> Bear River                        0.0        0.0        0.0        0.0
#> Feather River              15291140.6 15308693.2 15311676.2 15311707.6
#> Yuba River                  2581927.8  2638604.8  2841731.4  3568829.8
#> Lower-mid Sacramento River        0.0        0.0        0.0        0.0
#> Yolo Bypass                       0.0        0.0        0.0        0.0
#> American River                    0.0        0.0        0.0        0.0
#> Lower Sacramento River            0.0        0.0        0.0        0.0
#> Calaveras River                   0.0        0.0        0.0        0.0
#> Cosumnes River                    0.0        0.0        0.0        0.0
#> Mokelumne River                   0.0        0.0        0.0        0.0
#> Merced River                      0.0        0.0        0.0        0.0
#> Stanislaus River               7818.6     7818.6     7818.6    88139.8
#> Tuolumne River                  376.8      376.8      376.8      533.8
#> San Joaquin River                 0.0        0.0        0.0        0.0
#>                                    18         19         20
#> Upper Sacramento River     18259447.7 20516584.7 21623423.0
#> Antelope Creek              4384853.0  5145172.6  5801181.4
#> Battle Creek                4911148.4  5581978.0  5912117.6
#> Bear Creek                        0.0        0.0        0.0
#> Big Chico Creek             4671566.4  5498830.8  6216948.8
#> Butte Creek                27927379.8 34289773.4 37290294.6
#> Clear Creek                 1349854.5  1518793.0  1595221.2
#> Cottonwood Creek             240806.6   292145.6   293056.2
#> Cow Creek                         0.0        0.0        0.0
#> Deer Creek                  3378420.2  3749097.2  4121281.4
#> Elder Creek                       0.0        0.0        0.0
#> Mill Creek                 11128034.4 12828313.0 14211106.2
#> Paynes Creek                      0.0        0.0        0.0
#> Stony Creek                       0.0        0.0        0.0
#> Thomes Creek                   5181.0     6405.6     6625.4
#> Upper-mid Sacramento River        0.0        0.0        0.0
#> Sutter Bypass                     0.0        0.0        0.0
#> Bear River                        0.0        0.0        0.0
#> Feather River              15311707.6 19629615.8 19700768.2
#> Yuba River                  4028588.6  5083503.0  5838641.6
#> Lower-mid Sacramento River        0.0        0.0        0.0
#> Yolo Bypass                       0.0        0.0        0.0
#> American River                    0.0        0.0        0.0
#> Lower Sacramento River            0.0        0.0        0.0
#> Calaveras River                   0.0        0.0        0.0
#> Cosumnes River                    0.0        0.0        0.0
#> Mokelumne River                   0.0        0.0        0.0
#> Merced River                      0.0        0.0        0.0
#> Stanislaus River              88139.8    88139.8    88139.8
#> Tuolumne River                  533.8      533.8      533.8
#> San Joaquin River                 0.0        0.0        0.0
#> 
#> $proportion_natural
#>                                    1         2         3         4         5
#> Upper Sacramento River     1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Antelope Creek             0.9983498 0.9942363 0.9944444 1.0000000 0.9966667
#> Battle Creek               0.9873798 0.9672387 0.9432049 0.8562500 0.9366197
#> Bear Creek                        NA        NA        NA        NA        NA
#> Big Chico Creek            0.9932584 0.9854015 0.9768786 0.9728507 0.9865772
#> Butte Creek                0.9974381 0.9949787 0.9923042 0.9938805 0.9970951
#> Clear Creek                0.9976048 0.9952229 0.9957447 0.9793814 1.0000000
#> Cottonwood Creek           0.9918699 1.0000000 1.0000000 0.9500000 0.9750000
#> Cow Creek                         NA        NA        NA        NA        NA
#> Deer Creek                 0.9763365 0.9566327 0.8989637 0.9356725 0.9435146
#> Elder Creek                       NA        NA        NA        NA        NA
#> Mill Creek                 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Paynes Creek                      NA        NA        NA        NA        NA
#> Stony Creek                       NA        NA        NA        NA        NA
#> Thomes Creek               1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Upper-mid Sacramento River        NA        NA        NA        NA        NA
#> Sutter Bypass                     NA        NA        NA        NA        NA
#> Bear River                        NA        NA        NA        NA        NA
#> Feather River              0.8557137 0.7529430 0.4353096 0.4152370 0.6957052
#> Yuba River                 0.5301915 0.3270677 0.2596525 0.3382643 0.4343096
#> Lower-mid Sacramento River        NA        NA        NA        NA        NA
#> Yolo Bypass                       NA        NA        NA        NA        NA
#> American River                    NA        NA        NA        NA        NA
#> Lower Sacramento River            NA        NA        NA        NA        NA
#> Calaveras River            1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Cosumnes River             1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Mokelumne River            1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Merced River               1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Stanislaus River           1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Tuolumne River             1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> San Joaquin River                 NA        NA        NA        NA        NA
#>                                    6         7         8         9         10
#> Upper Sacramento River     1.0000000 1.0000000 1.0000000 1.0000000 1.00000000
#> Antelope Creek             1.0000000 1.0000000 1.0000000 0.9890110 1.00000000
#> Battle Creek               0.9691290 0.9777961 0.9696000 0.8730159 0.79470199
#> Bear Creek                        NA        NA        NA        NA         NA
#> Big Chico Creek            0.9862637 0.9782609 0.9894180 0.9222222 0.89655172
#> Butte Creek                0.9955086 0.9962704 0.9961390 0.9890378 0.98305085
#> Clear Creek                0.9910515 0.9983471 1.0000000 0.9354839 0.96363636
#> Cottonwood Creek           1.0000000 0.9870130 0.9200000 1.0000000 0.80000000
#> Cow Creek                         NA        NA        NA        NA         NA
#> Deer Creek                 0.9363905 0.9616909 0.9540481 0.8693182 0.64814815
#> Elder Creek                       NA        NA        NA        NA         NA
#> Mill Creek                 1.0000000 1.0000000 1.0000000 1.0000000 1.00000000
#> Paynes Creek                      NA        NA        NA        NA         NA
#> Stony Creek                       NA        NA        NA        NA         NA
#> Thomes Creek               1.0000000 1.0000000 1.0000000 1.0000000 1.00000000
#> Upper-mid Sacramento River        NA        NA        NA        NA         NA
#> Sutter Bypass                     NA        NA        NA        NA         NA
#> Bear River                        NA        NA        NA        NA         NA
#> Feather River              0.7089322 0.7388622 0.5447982 0.1472059 0.09185132
#> Yuba River                 0.3330702 0.3970117 0.3388042 0.1529889 0.10821918
#> Lower-mid Sacramento River        NA        NA        NA        NA         NA
#> Yolo Bypass                       NA        NA        NA        NA         NA
#> American River                    NA        NA        NA        NA         NA
#> Lower Sacramento River            NA        NA        NA        NA         NA
#> Calaveras River            1.0000000 1.0000000 1.0000000 1.0000000 1.00000000
#> Cosumnes River             1.0000000 1.0000000 1.0000000 1.0000000 1.00000000
#> Mokelumne River            1.0000000 1.0000000 1.0000000 1.0000000 1.00000000
#> Merced River               1.0000000 1.0000000 1.0000000 1.0000000 1.00000000
#> Stanislaus River           1.0000000 1.0000000 1.0000000 1.0000000 1.00000000
#> Tuolumne River             1.0000000 1.0000000 1.0000000 1.0000000 1.00000000
#> San Joaquin River                 NA        NA        NA        NA         NA
#>                                     11         12          13          14
#> Upper Sacramento River     1.000000000 1.00000000 1.000000000 1.000000000
#> Antelope Creek             1.000000000 1.00000000 1.000000000 0.875000000
#> Battle Creek               0.593023256 0.71764706 0.578125000 0.421052632
#> Bear Creek                          NA         NA          NA          NA
#> Big Chico Creek            0.684210526 0.81481481 0.750000000 0.750000000
#> Butte Creek                0.961098398 0.96259352 0.962025316 0.923857868
#> Clear Creek                0.935483871 0.96969697 0.823529412 1.000000000
#> Cottonwood Creek           1.000000000 0.50000000 1.000000000 0.500000000
#> Cow Creek                           NA         NA          NA          NA
#> Deer Creek                 0.554216867 0.60439560 0.571428571 0.288888889
#> Elder Creek                         NA         NA          NA          NA
#> Mill Creek                 1.000000000 1.00000000 1.000000000 1.000000000
#> Paynes Creek                        NA         NA          NA          NA
#> Stony Creek                         NA         NA          NA          NA
#> Thomes Creek                        NA         NA 1.000000000 1.000000000
#> Upper-mid Sacramento River          NA         NA          NA          NA
#> Sutter Bypass                       NA         NA          NA          NA
#> Bear River                          NA         NA          NA          NA
#> Feather River              0.002240896 0.00397966 0.002713704 0.002840176
#> Yuba River                 0.047151277 0.04768786 0.038284839 0.018315018
#> Lower-mid Sacramento River          NA         NA          NA          NA
#> Yolo Bypass                         NA         NA          NA          NA
#> American River                      NA         NA          NA          NA
#> Lower Sacramento River              NA         NA          NA          NA
#> Calaveras River            1.000000000 1.00000000 1.000000000          NA
#> Cosumnes River             1.000000000 1.00000000 1.000000000 1.000000000
#> Mokelumne River            1.000000000 1.00000000 1.000000000 1.000000000
#> Merced River               1.000000000 1.00000000          NA 1.000000000
#> Stanislaus River           1.000000000 1.00000000 1.000000000 1.000000000
#> Tuolumne River             1.000000000 1.00000000          NA 1.000000000
#> San Joaquin River                   NA         NA          NA          NA
#>                                    15         16         17         18
#> Upper Sacramento River     1.00000000 1.00000000 1.00000000 1.00000000
#> Antelope Creek             0.95454545 0.95454545 1.00000000 0.99212598
#> Battle Creek               0.80272109 0.88257576 0.96576577 0.98582817
#> Bear Creek                         NA         NA         NA         NA
#> Big Chico Creek            1.00000000 0.83333333 0.91891892 0.97196262
#> Butte Creek                0.97976391 0.98071867 0.99456790 0.99865643
#> Clear Creek                0.86363636 0.95238095 0.97029703 0.98127341
#> Cottonwood Creek           0.50000000 0.92857143 1.00000000 1.00000000
#> Cow Creek                          NA         NA         NA         NA
#> Deer Creek                 0.65486726 0.81632653 0.93092105 0.96734059
#> Elder Creek                        NA         NA         NA         NA
#> Mill Creek                 1.00000000 1.00000000 1.00000000 1.00000000
#> Paynes Creek                       NA         NA         NA         NA
#> Stony Creek                        NA         NA         NA         NA
#> Thomes Creek               1.00000000 1.00000000 1.00000000 1.00000000
#> Upper-mid Sacramento River         NA         NA         NA         NA
#> Sutter Bypass                      NA         NA         NA         NA
#> Bear River                         NA         NA         NA         NA
#> Feather River              0.02355580 0.04971537 0.04052802 0.03276784
#> Yuba River                 0.02136182 0.03120760 0.06957929 0.17493113
#> Lower-mid Sacramento River         NA         NA         NA         NA
#> Yolo Bypass                        NA         NA         NA         NA
#> American River                     NA         NA         NA         NA
#> Lower Sacramento River             NA         NA         NA         NA
#> Calaveras River            1.00000000 1.00000000 1.00000000 1.00000000
#> Cosumnes River             1.00000000 1.00000000 1.00000000 1.00000000
#> Mokelumne River            1.00000000 1.00000000 1.00000000 1.00000000
#> Merced River               1.00000000 1.00000000 1.00000000 1.00000000
#> Stanislaus River           1.00000000 1.00000000 1.00000000 1.00000000
#> Tuolumne River             1.00000000 1.00000000 1.00000000 1.00000000
#> San Joaquin River                  NA         NA         NA         NA
#>                                    19         20
#> Upper Sacramento River     1.00000000 1.00000000
#> Antelope Creek             1.00000000 0.99785867
#> Battle Creek               0.97595474 0.98067287
#> Bear Creek                         NA         NA
#> Big Chico Creek            0.98823529 0.99557522
#> Butte Creek                0.99689882 0.99849995
#> Clear Creek                1.00000000 0.99497487
#> Cottonwood Creek           1.00000000 0.98275862
#> Cow Creek                          NA         NA
#> Deer Creek                 0.96912900 0.97536567
#> Elder Creek                        NA         NA
#> Mill Creek                 1.00000000 1.00000000
#> Paynes Creek                       NA         NA
#> Stony Creek                        NA         NA
#> Thomes Creek               1.00000000 1.00000000
#> Upper-mid Sacramento River         NA         NA
#> Sutter Bypass                      NA         NA
#> Bear River                         NA         NA
#> Feather River              0.02148847 0.01678108
#> Yuba River                 0.23149016 0.41259321
#> Lower-mid Sacramento River         NA         NA
#> Yolo Bypass                        NA         NA
#> American River                     NA         NA
#> Lower Sacramento River             NA         NA
#> Calaveras River            1.00000000 1.00000000
#> Cosumnes River             1.00000000 1.00000000
#> Mokelumne River            1.00000000 1.00000000
#> Merced River               1.00000000 1.00000000
#> Stanislaus River           1.00000000 1.00000000
#> Tuolumne River             1.00000000 1.00000000
#> San Joaquin River                  NA         NA
#> 
#> $viability_metrics
#>       [,1] [,2] [,3] [,4]
#>  [1,]    0    2    2    0
#>  [2,]    0    0    1    0
#>  [3,]    0    0    0    0
#>  [4,]    0    0    0    0
#>  [5,]    0    0    0    0
#>  [6,]    0    0    1    0
#>  [7,]    0    1    1    0
#>  [8,]    0    0    1    0
#>  [9,]    0    0    1    0
#> [10,]    0    0    0    0
#> [11,]    0    0    0    0
#> [12,]    0    0    0    0
#> [13,]    0    0    0    0
#> [14,]    0    0    0    0
#> [15,]    0    0    0    0
#> [16,]    0    0    0    0
#> [17,]    0    0    0    0
#> [18,]    0    0    0    0
#> [19,]    0    1    0    0
#>